# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: API de Gestión de Alojamientos
  version: 1.0.0
  description: API para usuarios, anfitriones, alojamientos, reservas y comentarios. Rutas y respuestas basadas en los requerimientos del proyecto.
servers:
  - url: http://localhost:8080
paths:
...
  /alojamientos:
    post:
      summary: Crear nuevo alojamiento (solo anfitrión)
      operationId: crearAlojamiento
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlojamientoCrear'
      responses:
        '201':
          description: Alojamiento creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlojamientoResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Usuario no es anfitrión
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflicto de negocio (ej. límite de imágenes)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Buscar alojamientos (filtros y paginación)
      operationId: buscarAlojamientos
      parameters:
        - name: ciudad
          in: query
          schema:
            type: string
        - name: fechaInicio
          in: query
          schema:
            type: string
            format: date
        - name: fechaFin
          in: query
          schema:
            type: string
            format: date
        - name: precioMin
          in: query
          schema:
            type: number
        - name: precioMax
          in: query
          schema:
            type: number
        - name: servicios
          in: query
          schema:
            type: string
          description: Comma-separated (ej. wifi,piscina)
        - name: pagina
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Lista paginada de alojamientos
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  pagina:
                    type: integer
                  size:
                    type: integer
                  resultados:
                    type: array
                    items:
                      $ref: '#/components/schemas/AlojamientoResponse'
        '400':
          description: Filtros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /alojamientos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: ID del alojamiento
    get:
      summary: Ver detalle completo de un alojamiento
      operationId: verAlojamiento
      responses:
        '200':
          description: Detalle del alojamiento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlojamientoResponse'
        '404':
          description: Alojamiento no encontrado o eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Editar alojamiento (solo anfitrión propietario)
      operationId: editarAlojamiento
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlojamientoEditar'
      responses:
        '200':
          description: Alojamiento actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlojamientoResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: No es el propietario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No existe el alojamiento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Eliminar alojamiento (soft delete) — solo si no hay reservas futuras
      operationId: eliminarAlojamiento
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Alojamiento marcado como eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          description: Tiene reservas futuras o no es propietario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No existe el alojamiento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /usuarios/{id}/alojamientos:
    get:
      summary: Listar alojamientos de un anfitrión
      operationId: listarAlojamientosAnfitrion
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de alojamientos del anfitrión
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlojamientoResponse'
        '403':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        mensaje:
          type: string
          example: "Operación exitosa"
    ErrorResponse:
      type: object
      properties:
        codigo:
          type: string
          example: "ERR_VALIDACION"
        mensaje:
          type: string
          example: "El email ya está registrado"

    # -- Alojamiento --
    AlojamientoCrear:
      type: object
      required:
        - titulo
        - descripcion
        - ubicacion
        - precioPorNoche
        - capacidad
      properties:
        titulo:
          type: string
          example: "Casa en la playa"
        descripcion:
          type: string
        ubicacion:
          type: object
          required: [ciudad, direccion]
          properties:
            ciudad:
              type: string
            direccion:
              type: string
            lat:
              type: number
            lng:
              type: number
        precioPorNoche:
          type: number
        capacidad:
          type: integer
        servicios:
          type: array
          items:
            type: string
            example: "wifi"
        imagenes:
          type: array
          items:
            type: string
            example: "/img1.jpg"
    AlojamientoEditar:
      type: object
      properties:
        titulo:
          type: string
        descripcion:
          type: string
        precioPorNoche:
          type: number
        capacidad:
          type: integer
        servicios:
          type: array
          items:
            type: string
    AlojamientoResponse:
      type: object
      properties:
        id:
          type: string
          example: "a_98765"
        anfitrionId:
          type: string
        titulo:
          type: string
        descripcion:
          type: string
        ubicacion:
          type: object
          properties:
            ciudad: { type: string }
            direccion: { type: string }
            lat: { type: number }
            lng: { type: number }
        precioPorNoche:
          type: number
        capacidad:
          type: integer
        servicios:
          type: array
          items:
            type: string
        imagenes:
          type: array
          items:
            type: string
        calificacionPromedio:
          type: number
          format: float
        estado:
          type: string
          enum: [ACTIVO, ELIMINADO, BORRADOR]
...